requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# LangChain and AI
langchain==0.1.0
langchain-community==0.0.10

# Vector database
qdrant-client==1.7.0

# Embeddings
sentence-transformers==2.2.2

# Document processing
pandas==2.1.4
python-docx==1.1.0
pypdf==3.17.4
openpyxl==3.1.2

# Environment variables
python-dotenv==1.0.0

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
